reorganize tests for datastrategies
