+++ tests/_app/runtime/debug/index.data
@@ -1,1 +1,1 @@
-a:1:{s:13:"5a9087422ffb1";a:9:{s:3:"tag";s:13:"5a9087422ffb1";s:3:"url";s:38:"http://localhost:8000/index.php/api/f1";s:4:"ajax";i:0;s:6:"method";s:3:"GET";s:2:"ip";s:9:"127.0.0.1";s:4:"time";d:1519421250.19304;s:10:"statusCode";i:200;s:8:"sqlCount";i:0;s:9:"mailCount";i:0;}}
+a:5:{s:13:"5a9087422ffb1";a:9:{s:3:"tag";s:13:"5a9087422ffb1";s:3:"url";s:38:"http://localhost:8000/index.php/api/f1";s:4:"ajax";i:0;s:6:"method";s:3:"GET";s:2:"ip";s:9:"127.0.0.1";s:4:"time";d:1519421250.19304;s:10:"statusCode";i:200;s:8:"sqlCount";i:0;s:9:"mailCount";i:0;}s:13:"5be17b92df357";a:10:{s:3:"tag";s:13:"5be17b92df357";s:3:"url";s:33:"http://localhost:8080/api/f1/form";s:4:"ajax";i:0;s:6:"method";s:3:"GET";s:2:"ip";s:9:"127.0.0.1";s:4:"time";d:1541503890.69226;s:10:"statusCode";i:401;s:8:"sqlCount";i:0;s:9:"mailCount";i:0;s:9:"mailFiles";a:0:{}}s:13:"5be17bb108f45";a:10:{s:3:"tag";s:13:"5be17bb108f45";s:3:"url";s:45:"http://localhost:8080/api/f1/form?accessToken";s:4:"ajax";i:0;s:6:"method";s:3:"GET";s:2:"ip";s:9:"127.0.0.1";s:4:"time";d:1541503921.033568;s:10:"statusCode";i:401;s:8:"sqlCount";i:1;s:9:"mailCount";i:0;s:9:"mailFiles";a:0:{}}s:13:"5be17bb61e9d6";a:10:{s:3:"tag";s:13:"5be17bb61e9d6";s:3:"url";s:86:"http://localhost:8080/api/f1/form?accessToken=1234567812345678123456781234567812345678";s:4:"ajax";i:0;s:6:"method";s:3:"GET";s:2:"ip";s:9:"127.0.0.1";s:4:"time";d:1541503926.122095;s:10:"statusCode";i:200;s:8:"sqlCount";i:7;s:9:"mailCount";i:0;s:9:"mailFiles";a:0:{}}s:13:"5be1ee372c4f7";a:10:{s:3:"tag";s:13:"5be1ee372c4f7";s:3:"url";s:82:"http://localhost:8080/index.php/api/f1/field/1/rule?created_by=1&expand=properties";s:4:"ajax";i:0;s:6:"method";s:3:"GET";s:2:"ip";s:9:"127.0.0.1";s:4:"time";d:1541533238.799941;s:10:"statusCode";i:200;s:8:"sqlCount";i:6;s:9:"mailCount";i:0;s:9:"mailFiles";a:0:{}}}
