+++ tests/api/form/SectionCest.php
@@ -280,6 +280,7 @@ class SectionCest extends \tecnocen\roa\test\AbstractResourceCest
      * @param  Example $example
      * @dataprovider deleteDataProvider
      * @depends fixtures
+     * @depends form\section\SectionFieldCest:delete
      * @before authToken
      */
     public function delete(ApiTester $I, Example $example)
